removed the idm storage interface, merging its methods into IStorage and IStorageQuery
