+++ manager/api/war/src/main/java/io/apiman/manager/api/war/WarCdiFactory.java
@@ -20,7 +20,6 @@ import io.apiman.common.plugin.PluginClassLoader;
 import io.apiman.common.plugin.PluginCoordinates;
 import io.apiman.common.util.ReflectionUtils;
 import io.apiman.manager.api.core.IApiKeyGenerator;
-import io.apiman.manager.api.core.IIdmStorage;
 import io.apiman.manager.api.core.IMetricsAccessor;
 import io.apiman.manager.api.core.IPluginRegistry;
 import io.apiman.manager.api.core.IServiceCatalog;
@@ -37,7 +36,6 @@ import io.apiman.manager.api.core.noop.NoOpMetricsAccessor;
 import io.apiman.manager.api.es.ESMetricsAccessor;
 import io.apiman.manager.api.es.EsStorage;
 import io.apiman.manager.api.jpa.JpaStorage;
-import io.apiman.manager.api.jpa.roles.JpaIdmStorage;
 import io.apiman.manager.api.security.ISecurityContext;
 import io.apiman.manager.api.security.impl.DefaultSecurityContext;
 import io.apiman.manager.api.security.impl.KeycloakSecurityContext;
@@ -175,23 +173,6 @@ public class WarCdiFactory {
                     config.getServiceCatalogProperties(), pluginRegistry);
         } catch (Throwable t) {
             throw new RuntimeException("Error or unknown service catalog type: " + config.getServiceCatalogType(), t); //$NON-NLS-1$
-        }
-    }
-
-    @Produces @ApplicationScoped
-    public static IIdmStorage provideIdmStorage(WarApiManagerConfig config, @New JpaIdmStorage jpaIdmStorage,
-            @New EsStorage esStorage, IPluginRegistry pluginRegistry) {
-        if ("jpa".equals(config.getStorageType())) { //$NON-NLS-1$
-            return jpaIdmStorage;
-        } else if ("es".equals(config.getStorageType())) { //$NON-NLS-1$
-            return initES(config, esStorage);
-        } else {
-            try {
-                return createCustomComponent(IIdmStorage.class, config.getIdmStorageType(),
-                        config.getIdmStorageProperties(), pluginRegistry);
-            } catch (Throwable t) {
-                throw new RuntimeException("Error or unknown IDM storage type: " + config.getIdmStorageType(), t); //$NON-NLS-1$
-            }
         }
     }
