+++ manager/test/api/pom.xml
@@ -53,10 +53,6 @@
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>apiman-manager-api-jpa-idm</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>${project.groupId}</groupId>
       <artifactId>apiman-manager-api-es</artifactId>
     </dependency>
     <dependency>
