+++ manager/test/api/src/main/java/io/apiman/manager/test/server/DefaultTestDataSeeder.java
@@ -16,7 +16,6 @@
 package io.apiman.manager.test.server;
 
 import io.apiman.manager.api.beans.idm.UserBean;
-import io.apiman.manager.api.core.IIdmStorage;
 import io.apiman.manager.api.core.IStorage;
 import io.apiman.manager.api.core.exceptions.StorageException;
 
@@ -28,11 +27,11 @@ import io.apiman.manager.api.core.exceptions.StorageException;
 public class DefaultTestDataSeeder implements ISeeder {
     
     /**
-     * @see io.apiman.manager.test.server.ISeeder#seed(io.apiman.manager.api.core.IIdmStorage, io.apiman.manager.api.core.IStorage)
+     * @see io.apiman.manager.test.server.ISeeder#seed(io.apiman.manager.api.core.IStorage)
      */
     @SuppressWarnings("nls")
     @Override
-    public void seed(IIdmStorage idmStorage, IStorage storage) throws StorageException {
+    public void seed(IStorage storage) throws StorageException {
         for (String [] userInfo : TestUsers.USERS) {
             UserBean userBean = new UserBean();
             userBean.setUsername(userInfo[0]);
@@ -41,7 +40,7 @@ public class DefaultTestDataSeeder implements ISeeder {
             if ("true".equals(System.getProperty("apiman.test.admin-user-only", "false")) && !userBean.getUsername().equals("admin")) {
                 continue;
             }
-            idmStorage.createUser(userBean);
+            storage.createUser(userBean);
         }
     }
