+++ pom.xml
@@ -206,11 +206,6 @@
       </dependency>
       <dependency>
         <groupId>${project.groupId}</groupId>
-        <artifactId>apiman-manager-api-jpa-idm</artifactId>
-        <version>${project.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>${project.groupId}</groupId>
         <artifactId>apiman-manager-api-es</artifactId>
         <version>${project.version}</version>
       </dependency>
